Split menu bar rendering from state
