+++ CMakeLists.txt
@@ -181,6 +181,7 @@ set(UI_NUKLEAR_GLFW_SOURCES
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/ui_menu_def.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/ui_menu_actions.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/ui_menubar.c
+    ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/ui_menubar_render.c
     ${CMAKE_CURRENT_SOURCE_DIR}/src/ui/ui_theme.c
 )
